refactor(src): reorganize utils folder structure to frontend-utils
* Move common frontend-specific utils into new frontend-utils directory
* Renamed utils.ts to fileListingUtils.ts for better clarity
* Update import paths across multiple files
