+++ coauthor-for-vs-code/src/commands/mergeCommands.ts
@@ -1,5 +1,5 @@
 import * as vscode from 'vscode';
-import { getConfig } from '../utils/commonUtils';
+import { getConfig } from '../frontend-utils/commonUtils';
 import { debug } from '../logger/logUtils';
 
 const CHANNEL = 'MergeCommands';
