+++ coauthor-for-vs-code/src/folderExplorer.ts
@@ -2,7 +2,7 @@
 import * as vscode from 'vscode';
 import * as path from 'path';
 import { debug, info, warn, error, initializeLogging } from './logger/logUtils';
-import { getConfig } from './utils/commonUtils';
+import { getConfig } from './frontend-utils/commonUtils';
 
 const CHANNEL = 'FolderExplorer';
 initializeLogging(CHANNEL);
