+++ coauthor-for-vs-code/src/frontend-utils/commonUtils.ts
