+++ coauthor-for-vs-code/src/frontend-utils/fileListingUtils.ts
@@ -1,8 +1,8 @@
 import * as vscode from 'vscode';
 import * as path from 'path';
-import { getConfig } from './utils/commonUtils';
-import { initializeLogging } from './logger/logUtils';
-import { getWorkspacePath } from './utils/fileUtils';
+import { getConfig } from './commonUtils';
+import { initializeLogging } from '../logger/logUtils';
+import { getWorkspacePath } from '../utils/fileUtils';
 
 const CHANNEL = 'Utils';
 initializeLogging(CHANNEL);
