+++ coauthor-for-vs-code/src/frontend-utils/terminal.ts
