+++ coauthor-for-vs-code/src/utils/promptUtils.ts
@@ -4,7 +4,7 @@ import { readFile } from './fileUtils';
 import * as yaml from 'yaml';
 import * as vscode from 'vscode';
 import * as path from 'path';
-import { getConfig } from './commonUtils';
+import { getConfig } from '../frontend-utils/commonUtils';
 
 const CHANNEL = 'PromptUtils';
 initializeLogging(CHANNEL);
